// nextjs/components/database/schemas/public/CampaignConfiguration.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

/** Identifier type for public.campaign_configuration */
export type Campaign_configurationName = string;

/** Identifier type for public.campaign_configuration */
export type Campaign_configurationCompany = string;

/** Represents the table public.campaign_configuration */
export default interface CampaignConfiguration {
  name: Campaign_configurationName;

  company: Campaign_configurationCompany;

  measureValueNames: string[];

  entityIds: string[];
}

/** Represents the initializer for the table public.campaign_configuration */
export interface CampaignConfigurationInitializer {
  name: Campaign_configurationName;

  company: Campaign_configurationCompany;

  measureValueNames: string[];

  entityIds: string[];
}

/** Represents the mutator for the table public.campaign_configuration */
export interface CampaignConfigurationMutator {
  name?: Campaign_configurationName;

  company?: Campaign_configurationCompany;

  measureValueNames?: string[];

  entityIds?: string[];
}